Refer to `p` for what it is (a Kernel method not a keyword)
For the sake of correctness :)

Side discussion: This also means that it _is_ possible to
override the p function and have it function as a paragraph
function.

Whether or not to do it is a discussion that should probably
be had in a different issue.

It's the question if taking away the beloved p from debugging
inside their component (making it available as `_p` or what not)
is a reasonable exchange for having a more direct mapping to the
HTML `<p>` tag and that people don't have to keep that one
exception mind.