Pragmatic fix for failing test
`inject_into_file` only works when a file is actually present,
the README seems to be generated nowhere so it seems to assume
that it is manually created - hence manually created it in the
dummy app to make the test run pass.

Probably not the _best_ solution as the generator seems to mainly
be used in the core repo itself for which the dummy app doesn't
qualify. Hence working with a different example app/folder might
be the preferred solution.